--- frontend-react/src/telemetry-provider.tsx
@@ -1,13 +1,0 @@
-import React, { PropsWithChildren, useEffect } from "react";
-
-import { ai } from "./TelemetryService";
-
-const TelemetryProvider = ({ children }: PropsWithChildren<{}>) => {
-    useEffect(() => {
-        ai.initialize();
-    }, []);
-
-    return <>children</>;
-};
-
-export default TelemetryProvider;
